move traits to mod
