+++ jolt-core/benches/binding.rs
@@ -5,7 +5,7 @@ use criterion::Criterion;
 use jolt_core::field::JoltField;
 use jolt_core::poly::compact_polynomial::CompactPolynomial;
 use jolt_core::poly::dense_mlpoly::DensePolynomial;
-use jolt_core::poly::multilinear_polynomial::{BindingOrder, PolynomialBinding};
+use jolt_core::poly::{BindingOrder, PolynomialBinding};
 use rayon::prelude::*;
 
 fn random_dense_coeffs<F: JoltField>(rng: &mut impl Rng, num_vars: usize) -> Vec<F> {
