+++ jolt-core/src/poly/identity_poly.rs
@@ -13,7 +13,7 @@ use crate::utils::lookup_bits::LookupBits;
 use crate::utils::math::Math;
 use crate::utils::uninterleave_bits;
 
-use super::multilinear_polynomial::{BindingOrder, PolynomialBinding, PolynomialEvaluation};
+use super::{BindingOrder, PolynomialBinding, PolynomialEvaluation};
 
 #[derive(Clone, Debug, Allocative)]
 pub struct IdentityPolynomial<F: JoltField> {
