+++ jolt-core/src/zkvm/instruction_lookups/ra_virtual.rs
@@ -5,12 +5,12 @@ use crate::{
     poly::{
         commitment::commitment_scheme::CommitmentScheme,
         eq_poly::EqPolynomial,
-        multilinear_polynomial::{BindingOrder, PolynomialBinding},
         opening_proof::{
             OpeningPoint, ProverOpeningAccumulator, SumcheckId, VerifierOpeningAccumulator,
             BIG_ENDIAN,
         },
         ra_poly::RaPolynomial,
+        BindingOrder, PolynomialBinding,
     },
     subprotocols::{mles_product_sum::compute_mles_product_sum, sumcheck::SumcheckInstance},
     transcripts::Transcript,
